[feat] 변경된 기능으로 로직 변경
